sonar critical for errorhandling
Cast one of the operands of this multiplication operation to a long
Sonar Link:
https://sonar.onap.org/project/issues?assignees=dthakur&id=org.onap.dmaap.messagerouter.msgrtr%3Amsgrtr&open=AV4-VaiX32hFUzlqc5u2&resolved=false&types=BUG
Location:
src/main/java/com/att/nsa/cambria/security/DMaaPAuthenticatorImpl.java
Line No-51

Issue-ID: DMAAP-364

+++ src/main/java/com/att/nsa/cambria/security/DMaaPAuthenticatorImpl.java
@@ -48,7 +48,7 @@ public class DMaaPAuthenticatorImpl<K extends NsaApiKey> implements DMaaPAuthent
 	// Setting timeout to a large value for testing purpose.
 	// private static final long kDefaultRequestTimeWindow = 1000 * 60 * 10; //
 	// 10 minutes
-	private static final long kDefaultRequestTimeWindow = 1000 * 60 * 10 * 10 * 10 * 10 * 10;
+	private static final long kDefaultRequestTimeWindow = (long)1000 * 60 * 10 * 10 * 10 * 10 * 10;
 
 	/**
 	 * Construct the security manager against an API key database
